multi_agent_scraper/pyproject.toml
[project]
name = "multi-agent-scraper"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.13.4",
    "google-generativeai>=0.8.5",
    "gspread>=6.2.0",
    "hatchling>=1.27.0",
    "oauth2client>=4.1.3",
    "openai-agents>=0.0.14",
    "requests>=2.32.3",
    "selenium>=4.32.0",
    "streamlit>=1.45.0",
]
